+++ packages/solid-1.0/src/useWallet.ts
@@ -1,14 +1,12 @@
 import { useStore } from "@nanostores/solid"
-import { Accessor, createEffect, onCleanup, onMount } from "solid-js"
+import { createEffect, onCleanup, onMount } from "solid-js"
 import {
   initStore,
   dispatchConnect,
   dispatchDisconnect,
   StoreProps,
-  WalletInfo,
-  AccountInfo,
-  Cluster,
 } from "@solana-wallets/core-1.0"
+import "@solana-mobile/wallet-adapter-mobile"
 
 import { createContextProvider } from "./context"
 
@@ -34,14 +32,14 @@ const [WalletProvider, _useWallet] = createContextProvider((props: StoreProps) =
     localStorageKey: props.localStorageKey,
     additionalWallets: props.additionalWallets,
   })
-  const wallets: Accessor<WalletInfo[]> = useStore($wallets)
-  const walletsByName: Accessor<Record<string, WalletInfo>> = useStore($walletsMap)
-  const connectedAccount: Accessor<AccountInfo | undefined> = useStore($connectedAccount)
-  const connected: Accessor<boolean> = useStore($isConnected)
-  const wallet: Accessor<WalletInfo | undefined> = useStore($wallet)
-  const env: Accessor<Cluster> = useStore($env)
-  const connecting: Accessor<boolean> = useStore($connecting)
-  const disconnecting: Accessor<boolean> = useStore($disconnecting)
+  const wallets = useStore($wallets)
+  const walletsByName = useStore($walletsMap)
+  const connectedAccount = useStore($connectedAccount)
+  const connected = useStore($isConnected)
+  const wallet = useStore($wallet)
+  const env = useStore($env)
+  const connecting = useStore($connecting)
+  const disconnecting = useStore($disconnecting)
 
   onMount(() => {
     const cleanup = initOnMount()