+++ packages/core-1.0/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solana-wallets/core-1.0
 
+## 0.8.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @solana-wallets/core@0.8.1
+
 ## 0.8.1
 
 ### Patch Changes