+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @solana-wallets/core
 
+## 0.8.1
+
+### Patch Changes
+
+- test dynamic types
+
 ## 0.8.0
 
 ### Minor Changes