+++ packages/solid-1.0/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solana-wallets/solid-1.0
 
+## 0.10.3
+
+### Patch Changes
+
+- test dynamic types
+  - @solana-wallets/core-1.0@0.8.2
+
 ## 0.10.2
 
 ### Patch Changes